labs/cdk/msk/requirements.txt
aws-cdk-lib==2.87.0
constructs>=10.0.0
aws-cdk.aws-msk-alpha>=2.87.0a0

